% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{cloglog_inv_deriv}
\alias{cloglog_inv_deriv}
\title{Derivate of inverse of cloglog link function}
\usage{
cloglog_inv_deriv(eta)
}
\value{
derivative of inverse cloglog(eta) wrt eta
}
\description{
Derivate of inverse of cloglog link function
}
