% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{get_time_quantiles}
\alias{get_time_quantiles}
\title{Computes times when survival function drops below given quantile values}
\usage{
get_time_quantiles(t, s_t, qs)
}
\arguments{
\item{t}{vector of time points}

\item{s_t}{corresponding vector of survival probabilites from KM estimate}

\item{qs}{quantiles}
}
\value{
Vector of time points based on quantiles
}
\description{
Computes times when survival function drops below given quantile values
}
