% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{hazard_function_cloglog.v}
\alias{hazard_function_cloglog.v}
\title{Hazard function at given times from I-spline parameters, cloglog link}
\usage{
hazard_function_cloglog.v(
  t,
  knots,
  degree,
  boundary_knots,
  coefficients,
  restriction = 0
)
}
\arguments{
\item{t}{vector of times}

\item{knots}{knot locations}

\item{degree}{spline degrees of freedom}

\item{boundary_knots}{boundary knot locations}

\item{coefficients}{estimated spline coefficients}
}
\value{
estimated hazard at given times
}
\description{
Hazard function at given times from I-spline parameters, cloglog link
}
