% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{hybrid_integration}
\alias{hybrid_integration}
\title{Combines rmutil::int and stats::integrate to quickly and accurately integrate steep one-dimensional functions with vector-valued upper limit}
\usage{
hybrid_integration(f, lower, upper, disconts = NULL, pivot = lower)
}
\arguments{
\item{f}{function}

\item{lower}{lower limit (one value)}

\item{upper}{upper limits (vector-valued)}

\item{disconts}{jump discontinuities (ignored if < lower or > pivot), increased accuracy if provided}

\item{pivot}{integration below pivot performed by Gauss-Konrad quadrature (accurate but expensive), above pivot by Romberg integration (fast but inaccurate for steep functions)}
}
\value{
vector of definite integral values
}
\description{
Combines rmutil::int and stats::integrate to quickly and accurately integrate steep one-dimensional functions with vector-valued upper limit
}
