% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{log_rank_numerator}
\alias{log_rank_numerator}
\title{Numerator of the two-sample log-rank test (observed-expected) based on spline parameters}
\usage{
log_rank_numerator(
  spline_params_all,
  spline_params_group1,
  restriction = 0,
  limit = spline_params_all$Max * 0.99
)
}
\arguments{
\item{spline_params_all}{spline parameters in the full population, containing knots, degrees of freedom, boundary knots and coefficients for both at-risk and event functions}

\item{spline_params_group1}{spline parameters in one of the groups, containing knots, degrees of freedom, boundary knots and coefficients for both at-risk and event functions}

\item{restriction}{pivot for hybrid integration (should be >0 if steep hazard observed near 0)}

\item{limit}{integration limit for log-rank test to prevent error due to extrapolation, a good choice is the maximum value prior to which we have events in both groups}
}
\value{
Numerator of log-rank test
}
\description{
Numerator of the two-sample log-rank test (observed-expected) based on spline parameters
}
