% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{squared_sum_influence}
\alias{squared_sum_influence}
\title{Sum of squared KM influence function values for a set of observations (for variance estimation) for a given time}
\usage{
squared_sum_influence(
  time,
  delta,
  spline_params,
  confint_time,
  restriction = 0,
  weights = rep(1, length(time))
)
}
\arguments{
\item{time}{vector of survival times}

\item{delta}{vector of censoring indicators}

\item{spline_params}{spline parameters, containing knots, degrees of freedom, boundary knots and coefficients for both at-risk and event functions}

\item{confint_time}{time we want to compute the influence function values for}

\item{restriction}{pivot for hybrid integration (should be >0 if steep hazard observed near 0)}

\item{weights}{vector of weights}
}
\value{
sum of squared influence function values
}
\description{
Sum of squared KM influence function values for a set of observations (for variance estimation) for a given time
}
