% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{squared_sum_influence_logrank}
\alias{squared_sum_influence_logrank}
\title{Sum of squared logrank (observed-expected) influence function values for a set of observations (for variance estimation in weighted logrank test)}
\usage{
squared_sum_influence_logrank(
  time,
  delta,
  treatment,
  spline_params,
  spline_params_treated,
  restriction = 0,
  weights = rep(1, length(time)),
  p_treated = 0.5,
  limit = spline_params$Max * 0.99
)
}
\arguments{
\item{time}{vector of survival times}

\item{delta}{vector of censoring indicators}

\item{treatment}{binary vector for treatment assignment}

\item{spline_params}{spline parameters, containing knots, degrees of freedom, boundary knots and coefficients for both at-risk and event functions}

\item{spline_params_treated}{spline parameters for treatment group, containing knots, degrees of freedom, boundary knots and coefficients for both at-risk and event functions}

\item{restriction}{pivot for hybrid integration (should be >0 if steep hazard observed near 0)}

\item{weights}{vector of weights}

\item{p_treated}{proportion of population that is treated in the weighted sample}

\item{limit}{integration limit for log-rank test to prevent error due to extrapolation, a good choice is the maximum value prior to which we have events in both groups}
}
\value{
sum of squared influence function values
}
\description{
Sum of squared logrank (observed-expected) influence function values for a set of observations (for variance estimation in weighted logrank test)
}
