% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{theoretical_KM_PS}
\alias{theoretical_KM_PS}
\title{Kaplan-Meier pseudo values derived from spline estimates and influence function}
\usage{
theoretical_KM_PS(spline_params, t_ps, time, delta, restriction = 0)
}
\arguments{
\item{spline_params}{spline parameters, containing knots, degrees of freedom, boundary knots and coefficients for both at-risk and event functions}

\item{t_ps}{time of interest}

\item{time}{vector of survival times}

\item{delta}{vector of censoring indicators}

\item{restriction}{pivot for hybrid integration (should be >0 if steep hazard observed near 0)}
}
\value{
pseudo values for each observation at time t_ps
}
\description{
Kaplan-Meier pseudo values derived from spline estimates and influence function
}
