% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{update_survival_splines_chunk}
\alias{update_survival_splines_chunk}
\title{Update risk and survival function splines for a given site in chunks, cloglog link}
\usage{
update_survival_splines_chunk(
  spline_params,
  time,
  delta,
  chunksize = 5,
  breakpoints = 5,
  knots_list = rep(length(spline_params$Event$Knots), ceiling(length(time)/chunksize)),
  degree = spline_params$Event$Degree,
  max_time = spline_params$Max,
  loud = TRUE,
  restriction = 0
)
}
\arguments{
\item{spline_params}{spline parameters, containing knots, degrees of freedom, boundary knots and coefficients for both at-risk and event functions}

\item{time}{vector of survival times}

\item{delta}{vector of censoring indicators}

\item{chunksize}{number of observations to update at a time (would not recommend above 10)}

\item{breakpoints}{number of breakpoints to use between each knot when adjusting spline}

\item{knots_list}{number of knots for each chunk (should have length ceiling(length(time)/chunksize) and be non-decreasing)}

\item{degree}{spline degrees of freedom}

\item{max_time}{time to estimate spline to}

\item{loud}{whether to print out progress in number of observations passed}

\item{restriction}{pivot for hybrid integration (should be >0 if steep hazard observed near 0)}
}
\value{
updated spline parameters for risk and survival functions
}
\description{
Update risk and survival function splines for a given site in chunks, cloglog link
}
