% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{update_survival_splines_weighted}
\alias{update_survival_splines_weighted}
\title{Update risk and survival function splines given survival observations with weights, cloglog link}
\usage{
update_survival_splines_weighted(
  spline_params,
  time,
  delta,
  weights = rep(1, length(time)),
  breakpoints = 5,
  knots = length(spline_params$Event$Knots),
  degree = spline_params$Event$Degree,
  confint_times = names(spline_params$Influence),
  max_time = spline_params$Max,
  restriction = 0
)
}
\arguments{
\item{spline_params}{spline parameters, containing knots, degrees of freedom, boundary knots and coefficients for both at-risk and event functions}

\item{time}{vector of survival times}

\item{delta}{vector of censoring indicators}

\item{weights}{vector of weights}

\item{breakpoints}{number of breakpoints to use between each knot when adjusting spline}

\item{knots}{number of knots, can be larger than previous but should not increase by more than 1-2 at a given site}

\item{degree}{spline degrees of freedom}

\item{confint_times}{times to store squared sum of influence functions for variance estimation in weighted case}

\item{max_time}{time to estimate spline to}

\item{restriction}{pivot for hybrid integration (should be >0 if steep hazard observed near 0)}
}
\value{
updated spline parameters for risk and survival functions
}
\description{
Update risk and survival function splines given survival observations with weights, cloglog link
}
