% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{variance_cloglog}
\alias{variance_cloglog}
\title{Variance of survival function at given times from I-spline parameters, cloglog link}
\usage{
variance_cloglog(t_k, spline_params, restriction = 0)
}
\arguments{
\item{t_k}{vector of times}

\item{spline_params}{spline parameters, containing knots, degrees of freedom, boundary knots and coefficients for both at-risk and event functions}

\item{restriction}{pivot for hybrid integration (should be >0 if steep hazard observed near 0)}
}
\value{
Variance of cloglog(S(t)) for each t in t_k
}
\description{
Variance of survival function at given times from I-spline parameters, cloglog link
}
